Update email address on about page/Home and navbar slight amendment made
The email on the about page has been updated, but when reducing the screen size the email goes off screen so there are sizing issues which need fixing therefore this task has been put on hold. On home page some text has been added for the user that the website is under maintenance.