update qualification page
I have updated the qualification section on the website, so I have added a lot of new qualifications to the qualification page. I have also removed the old maintenance banner on the top of the nav_bar and added a new maintenance banner to the home page of the website. I have also updated the folder structure of the images folder due to that update the nav bar logo link was update, images link in the project page were update and some updates on the multisync project page due to a lot of missing information.